feat(getInitialTravelTitle): 제목 초기값에서 사용자 닉네임 삭제
- 사용자 닉네임을 삭제하면서 훅이 아니게되어 함수이름을 useInitialTravelTitle에서 getInitialTravelTitle로 수정하고 util 폴더로 이동 시켰습니다.
